% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defi.R
\name{defi_tvl_history}
\alias{defi_tvl_history}
\title{Fetch historical total TVL across all DeFi protocols}
\usage{
defi_tvl_history()
}
\value{
tibble: date (Date), tvl (numeric)
}
\description{
Fetch historical total TVL across all DeFi protocols
}
